requirements.txt
absl-py==0.7.1
addict==2.2.1
albumentations==0.4.3
apex==0.1
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
catfish==1.4.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
conda==4.3.16
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
faiss==1.5.3
faiss-cpu==1.6.1
faiss-gpu==1.6.1
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
graphviz==0.8.4
grpcio==1.21.1
h5py==2.9.0
horovod==0.16.4
httplib2==0.9.2
idna==2.8
imagecorruptions==1.1.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
macaroonbakery==1.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
menulibre==2.2.0
mistune==0.8.4
mmcv==0.2.14
-e git+https://github.com/open-mmlab/mmdetection.git@b7894cbdcbe114e3e9efdd1a6a229419a552c807#egg=mmdet
mugshot==0.4.0
mxnet-cu100==1.4.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.14.6
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.3
ptyprocess==0.6.0
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.4.2
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.1
pyRFC3339==1.0
pyrsistent==0.15.3
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-debianbts==2.7.2
python-magic==0.4.16
pytorch-ignite==0.1.2
pytz==2019.1
pyxdg==0.25
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sgt-launcher==0.2.4
simplejson==3.13.2
six==1.12.0
sklearn==0.0
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thop==0.0.31.post1910280903
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.41.0
traitlets==4.3.2
typing==3.7.4
ubuntu-drivers-common==0.0.0
urllib3==1.25.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wrapt==1.11.2
xcffib==0.5.1
xkit==0.0.0
yacs==0.1.6
zope.interface==4.3.2